Rename to identifier and add test
