+++ packages/random/src/index.ts
@@ -1,6 +1,6 @@
 export * from './alpha.js';
 export * from './array.js';
+export * from './identifier.js';
 export * from './integer.js';
 export * from './sequence.js';
 export * from './uuid.js';
-export * from './wrapper.js';
